% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_names}
\alias{add_names}
\title{Add Names to Fields}
\usage{
add_names(fields_list, codierung_all)
}
\arguments{
\item{fields_list}{List of field data}
}
\value{
List of field data with added names
}
\description{
Add Names to Fields
}
