% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregator}
\alias{aggregator}
\title{Aggregate Field Data}
\usage{
aggregator(intersected, years, crop_codes, display_names)
}
\arguments{
\item{intersected}{sf object containing intersected fields}

\item{years}{Numeric vector of years to process}

\item{crop_codes}{crop codes for aggregaton}

\item{display_names}{names corresponding to the codes}
}
\value{
sf object with added aggregated columns
}
\description{
Aggregates field data by adding classified crop names for each year
}
