% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{Crop Rotation Visualization Server}
\usage{
app_server(input, output, session, app_data)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments}
}
\description{
Crop Rotation Visualization Server
}
\keyword{internal}
