% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classify_crops}
\alias{classify_crops}
\title{Classify Crops}
\usage{
classify_crops(nc_codes, names, crop_codes, display_names)
}
\arguments{
\item{nc_codes}{Numeric vector of crop codes}

\item{names}{Character vector of crop names}
}
\value{
Character vector of classified crop names
}
\description{
Classifies crops based on numeric codes and names using a lookup table
}
