% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_thin_polygons}
\alias{clean_thin_polygons}
\title{Clean Thin Polygons}
\usage{
clean_thin_polygons(sf_data, min_width = 1)
}
\arguments{
\item{sf_data}{SF object containing polygons}

\item{min_width}{Minimum width in meters}
}
\value{
SF object with cleaned polygons
}
\description{
Clean Thin Polygons
}
