% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_hex_color}
\alias{generate_hex_color}
\title{Function to Generate a Random Hex Color}
\usage{
generate_hex_color()
}
\value{
A hex color code representing a randomly generated color
}
\description{
Function to Generate a Random Hex Color
}
\examples{
Generate a single random color
random_color <- generate_hex_color()
print(random_color)

Generate a single random color
random_color <- generate_hex_color()
print(random_color)

Generate multiple random colors
multiple_colors <- replicate(5, generate_hex_color())
print(multiple_colors)

}
