% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{intersect_fields}
\alias{intersect_fields}
\title{Intersect Multiple Field Layers}
\usage{
intersect_fields(fields_list, max_area = 20000 * 1e+06)
}
\arguments{
\item{fields_list}{List of sf objects to intersect}
}
\value{
sf object with intersected fields
}
\description{
Intersects multiple spatial layers while maintaining CRS consistency
}
