% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{intersect_fields_simple}
\alias{intersect_fields_simple}
\title{Simple Field Intersection}
\usage{
intersect_fields_simple(fields_list, max_area = 20000 * 1e+06)
}
\arguments{
\item{fields_list}{List of sf objects to intersect}
}
\value{
sf object with intersected fields
}
\description{
Performs a simplified intersection of multiple spatial layers
}
