% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{intersecting_check_spatial}
\alias{intersecting_check_spatial}
\title{Check Spatial Intersections}
\usage{
intersecting_check_spatial(input_list, intersection)
}
\arguments{
\item{input_list}{List of sf objects to check}

\item{intersection}{sf object representing the intersection}
}
\value{
sf object with intersection results
}
\description{
Performs spatial intersection checking using a tiled approach for large datasets
}
